% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_conversion.R
\name{bilirubin_to_SI}
\alias{bilirubin_to_SI}
\title{Bilirubin unit converter (mg/dl -> µmol/l)}
\usage{
bilirubin_to_SI(bili)
}
\arguments{
\item{bili}{numeric vector of bilirubin levels (mg/dl)}
}
\value{
numeric vector of bilirubin levels in µmol/l
}
\description{
A vectorised function to convert serum bilirubin levels from mg/dl to µmol/l
}
\examples{
bilirubin_to_SI(bili = 3.1)
}
