% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_conversion.R
\name{BUN_to_urea}
\alias{BUN_to_urea}
\title{Convert BUN to urea}
\usage{
BUN_to_urea(BUN, units = "SI")
}
\arguments{
\item{BUN}{numeric vector of blood urea nitrogen levels (mg/dl)}

\item{units}{units for urea ("SI" for mmol/l, "US" for mg/dl)}
}
\value{
numeric vector of urea levels
}
\description{
A vectorised function to convert blood urea nitrogen (BUN) to urea.
The default unit for urea is mmol/l but this can be changed to mg/dl by
setting the optional units parameter to "US"
}
\examples{
BUN_to_urea(8.0)
}
