% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_conversion.R
\name{creatinine_to_SI}
\alias{creatinine_to_SI}
\title{Creatinine unit converter (mg/dl -> mcmol/l)}
\usage{
creatinine_to_SI(creat)
}
\arguments{
\item{creat}{numeric vector of creatinine levels (mg/dl)}
}
\value{
numeric vector of creatinine levels in mcmol/l
}
\description{
A vectorised function to convert serum creatinine levels from mg/dl to mcmol/l
}
\examples{
creatinine_to_SI(creat = 2.0)
}
