% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egfr.R
\name{ibw}
\alias{ibw}
\title{Ideal body weight}
\usage{
ibw(height, sex)
}
\arguments{
\item{height}{numeric vector of heights in cm}

\item{sex}{character vector of sex ("F" for female or "M" for male)}
}
\value{
numeric vector of ideal body weights in kg
}
\description{
A vectorised function to calculate adult ideal body weight based on height and sex.
This function assumes ideal BMI of 21.5 for females and 23 for males.
}
\examples{
ibw(height = 183, sex = "M")
}
