% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kidney.donors}
\alias{kidney.donors}
\title{Simulated dataset of donors to illustrate KDRI vignette.}
\format{A data frame with 4 rows and 7 variables:
\describe{
  \item{Donor.Age}{donor age in years}
  \item{Donor.Height}{donor height in cm}
  \item{Donor.Hypertension}{donor hypertension}
  \item{Donor.Sex}{donor sex}
  \item{Donor.CMV}{donor CMV status}
  \item{Donor.GFR}{donor GFR}
  \item{Donor.Hospital_Stay}{donor hospital stay in days}
  ...
}}
\usage{
kidney.donors
}
\description{
A simulated dataset of kidney donors to illustrate the KDRI vignette
}
\keyword{datasets}
