% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{liver.pts}
\alias{liver.pts}
\title{Simulated dataset to illustrate MELD calculator vignette.}
\format{A data frame with 4 rows and 6 variables:
\describe{
  \item{Patient.Age}{patient age in years}
  \item{Patient.INR}{Patient INR}
  \item{Patient.Bilirubin}{Patient serum bilirubin in µmol/l}
  \item{Patient.Creatinine}{serum serum creatinine in µmol/l}
  \item{Patient.Sodium}{Patient serum sodium in mmol/l}
  \item{Patient.Dialysed}{Whether patient dialysed (1 = yes, 0 = no)}
  ...
}}
\usage{
liver.pts
}
\description{
A simulated dataset of liver patient characteristics for the MELD vignette
}
\keyword{datasets}
