% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mismatches}
\alias{mismatches}
\title{Simulated dataset to illustrate mismatches for HLA vignette.}
\format{A data frame with 4 rows and 5 variables:
\describe{
  \item{HLA.A.MM}{HLA A mismatch}
  \item{HLA.B.MM}{HLA B mismatch}
  \item{HLA.DR.MM}{HLA DR mismatch}
  \item{HLA.MM}{HLA mismatch as string}
  \item{HLA.MM.s}{HLA mismatch as string with separators}
  ...
}}
\usage{
mismatches
}
\description{
A simulated dataset of HLA mismatches
}
\keyword{datasets}
