% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liver.R
\name{peld}
\alias{peld}
\title{PELD score}
\usage{
peld(INR, bili, albumin, listing_age, growth_failure, units = "SI")
}
\arguments{
\item{INR}{INR}

\item{bili}{serum biliruin (µmol/l)}

\item{albumin}{serum albumin}

\item{listing_age}{age at the time of listing (years; integer or decimal)}

\item{growth_failure}{whether there is growth failure (1 = yes, 0 = no)}

\item{units}{units used for bilirubin ("SI" for µmol/l (default), "US" for mg/dl)}
}
\value{
numeric vector of PELD scores
}
\description{
A vectorised function to generate a PELD score for paediatric liver transplant candidates.
The default unit for bilirubin is µmol/l but this can be changed to mg/dl by setting the
optional units parameter to "US".
}
\examples{
peld(INR = 2, bili = 54, albumin = 25, listing_age = 2, growth_failure = 1)
peld(INR = 2, bili = 3.1, albumin = 25, listing_age = 2, growth_failure = 1, units = "US")
}
