% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{results_US}
\alias{results_US}
\title{Simulated dataset to illustrate eGFR calculator vignette.}
\format{A data frame with 4 rows and 6 variables:
\describe{
  \item{Creatinine}{serum creatinine in mg/dl}
  \item{Sex}{Patient sex, M or F}
  \item{Age}{patient age in years}
  \item{Ethnicity}{Patient ethnicity}
  \item{Weight}{Patient weight in kg}
  \item{Height}{Patient height in cm}
  ...
}}
\usage{
results_US
}
\description{
A simulated dataset of patient characteristics and creatinine levels
}
\keyword{datasets}
