% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{serial.results}
\alias{serial.results}
\title{Simulated dataset to illustrate serial results eGFR calculator vignette.}
\format{A data frame with 4 rows and 6 variables:
\describe{
  \item{Age}{patient age in years}
  \item{Sex}{Patient sex, M or F}
  \item{Ethnicity}{Patient ethnicity}
  \item{Creatinine_1yr}{serum creatinine at 1 year, in µmol/l}
  \item{Creatinine_5yr}{serum creatinine at 5 years, in µmol/l}

  ...
}}
\usage{
serial.results
}
\description{
A simulated dataset of patient characteristics and creatinine levels
}
\keyword{datasets}
