% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukkri.R
\name{ukkdri}
\alias{ukkdri}
\title{Calculate UK Kidney Donor Risk Index (NHSBT, 2019 version)}
\usage{
ukkdri(age, height, htn, female, cmv, gfr, hdays)
}
\arguments{
\item{age}{numeric vector of donor age in years}

\item{height}{numeric vector of donor height in cm}

\item{htn}{numeric vector of whether donor history of hypertension (1 = yes, 0 = no)}

\item{female}{character vector of donor sex ("F" = female, "M" = male)}

\item{cmv}{numeric vector of whether donor CMV IgG positive (1 = yes, 0 = no)}

\item{gfr}{numeric vector of donor eGFR at time of donation}

\item{hdays}{numeric vector of number of days donor in hospital before donation}
}
\value{
numeric vector of UK Kidney Donor Risk Index values (2019 version)
}
\description{
A vectorised function to calculate the UK Kidney Donor Risk Index as used in the new
national kidney matching scheme implemented in September 2019.
}
\examples{
ukkdri(50, 170, 1, 1, 0, 90, 2)
}
