% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_conversion.R
\name{urea_to_bun}
\alias{urea_to_bun}
\title{Convert urea to BUN}
\usage{
urea_to_bun(urea, units = "SI")
}
\arguments{
\item{urea}{numeric vector of urea levels (mmol/l by default)}

\item{units}{units for urea ("SI" for mmol/l, "US" for mg/dl)}
}
\value{
numeric vector of blood urea nitrogen (BUN) levels in mg/dl
}
\description{
A vectorised function to convert urea to blood urea nitrogen (BUN),
By default the urea is measured in mmol/l but this can be changed to mg/dl by
setting the optional units parameter to "US"
}
\examples{
urea_to_bun(5.4)
}
