% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uskdri.R
\name{uskdri_US}
\alias{uskdri_US}
\title{US Kidney Donor Risk Index (US units)}
\usage{
uskdri_US(age, height, weight, eth, htn, dm, cva, creat, hcv, dcd)
}
\arguments{
\item{age}{numeric vector of donor ages in years}

\item{height}{numeric vector of donor heights in cm}

\item{weight}{numeric vector of donor weights in kg}

\item{eth}{character string vector of donor ethnicity, "black" or "non-black"}

\item{htn}{numeric vector of donor hypertension history (1 = yes, 0 = no)}

\item{dm}{numeric vector of donor diabetes history (1 = yes, 0 = no)}

\item{cva}{numeric vector of whether donor death due to CVA (1 = yes, 0 = no)}

\item{creat}{numeric vector of donor serum creatinine (mg/dl)}

\item{hcv}{numeric vector of donor hepatitis C history (1 = yes, 0 = no)}

\item{dcd}{numeric vector of type of donor (1 = DCD, 0 = DBD)}
}
\value{
numeric vector of US KDRI values
}
\description{
Wrapper function for the uskdri() vectorised function to calculate US Kidney Donor Risk Index
as published by UNOS but using mg/dl as the units for creatinine.
}
\examples{
uskdri(age = 40, height = 170, weight = 80, eth = "non-black", htn = 0, dm = 0,
    cva = 0, creat = 1.4, hcv = 0, dcd = 0) # 1.09
}
