% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uskdri.R
\name{uskdri}
\alias{uskdri}
\title{US Kidney Donor Risk Index}
\usage{
uskdri(age, height, weight, eth, htn, dm, cva, creat, hcv, dcd,
  scaling = 1, units = "SI")
}
\arguments{
\item{age}{numeric vector of donor ages in years}

\item{height}{numeric vector of donor heights in cm}

\item{weight}{numeric vector of donor weights in kg}

\item{eth}{character string vector of donor ethnicity, "black" or "non-black"}

\item{htn}{numeric vector of donor hypertension history (1 = yes, 0 = no)}

\item{dm}{numeric vector of donor diabetes history (1 = yes, 0 = no)}

\item{cva}{numeric vector of whether donor death due to CVA (1 = yes, 0 = no)}

\item{creat}{numeric vector of donor serum creatinine (µmol/l)}

\item{hcv}{numeric vector of donor hepatitis C history (1 = yes, 0 = no)}

\item{dcd}{numeric vector of type of donor (1 = DCD, 0 = DBD)}

\item{scaling}{single numeric value for OPTN scaling factor (optional, defaults to 1)}

\item{units}{single string value to indicate creatinine units ("SI" for µmol/l, "US" for mg/dl)}
}
\value{
numeric vector of US KDRI values
}
\description{
Vectorised function to calculate US Kidney Donor Risk Index as published by UNOS. Please note
that this function uses creatinine measured in µmol/l by default, but can be changed to
mg/dl if the optional units parameter is set to "US".
}
\details{
The KDRI is normalised by a scaling factor based on the median KDRI in the previous year. For 2018, this was
approximately 1.250609 as is published on the OPTN website. The scaling parameter in this function defaults to 1,
so can be left out to calculate the KDRI without scaling.
}
\examples{
# with creatinine in µmol/l (units = "SI" can be omitted)
uskdri(age = 40, height = 170, weight = 80, eth = "non-black", htn = 0, dm = 0,
    cva = 0, creat = 120, hcv = 0, dcd = 0, scaling = 1.250609, units = "SI")

# with creatinine in mg/dl
uskdri(age = 40, height = 170, weight = 80, eth = "non-black", htn = 0, dm = 0,
    cva = 0, creat = 1.4, hcv = 0, dcd = 0, scaling = 1.250609, units = "US")
}
